third_party/WebKit/LayoutTests/accessibility/table-header-column-row-expected.txt

scope test
col head	row head	col head
col head	data	row head
row head	data	col head
row header and column header (1)
data	row head	column head
column head	column head	data
row head	data	row head
row header and column header (2)
row head	data	data
column head	column head	column head
data	data	data
End of test

This tests that AXRoles for header cells are assigned.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: AXWebArea AXValue: 
    AXRole: AXTable AXValue: 
        AXRole: AXRow AXValue: 
            AXRole: AXColumnHeader AXValue: 
                AXRole: AXStaticText AXValue: col head
                    AXRole: AXInlineTextBox AXValue: col head
            AXRole: AXRowHeader AXValue: 
                AXRole: AXStaticText AXValue: row head
                    AXRole: AXInlineTextBox AXValue: row head
            AXRole: AXColumnHeader AXValue: 
                AXRole: AXStaticText AXValue: col head
                    AXRole: AXInlineTextBox AXValue: col head
        AXRole: AXRow AXValue: 
            AXRole: AXColumnHeader AXValue: 
                AXRole: AXStaticText AXValue: col head
                    AXRole: AXInlineTextBox AXValue: col head
            AXRole: AXCell AXValue: 
                AXRole: AXStaticText AXValue: data
                    AXRole: AXInlineTextBox AXValue: data
            AXRole: AXRowHeader AXValue: 
                AXRole: AXStaticText AXValue: row head
                    AXRole: AXInlineTextBox AXValue: row head
        AXRole: AXRow AXValue: 
            AXRole: AXRowHeader AXValue: 
                AXRole: AXStaticText AXValue: row head
                    AXRole: AXInlineTextBox AXValue: row head
            AXRole: AXCell AXValue: 
                AXRole: AXStaticText AXValue: data
                    AXRole: AXInlineTextBox AXValue: data
            AXRole: AXColumnHeader AXValue: 
                AXRole: AXStaticText AXValue: col head
                    AXRole: AXInlineTextBox AXValue: col head
    AXRole: AXTable AXValue: 
        AXRole: AXRow AXValue: 
            AXRole: AXCell AXValue: 
                AXRole: AXStaticText AXValue: data
                    AXRole: AXInlineTextBox AXValue: data
            AXRole: AXRowHeader AXValue: 
                AXRole: AXStaticText AXValue: row head
                    AXRole: AXInlineTextBox AXValue: row head
            AXRole: AXColumnHeader AXValue: 
                AXRole: AXStaticText AXValue: column head
                    AXRole: AXInlineTextBox AXValue: column head
        AXRole: AXRow AXValue: 
            AXRole: AXColumnHeader AXValue: 
                AXRole: AXStaticText AXValue: column head
                    AXRole: AXInlineTextBox AXValue: column head
            AXRole: AXColumnHeader AXValue: 
                AXRole: AXStaticText AXValue: column head
                    AXRole: AXInlineTextBox AXValue: column head
            AXRole: AXCell AXValue: 
                AXRole: AXStaticText AXValue: data
                    AXRole: AXInlineTextBox AXValue: data
        AXRole: AXRow AXValue: 
            AXRole: AXRowHeader AXValue: 
                AXRole: AXStaticText AXValue: row head
                    AXRole: AXInlineTextBox AXValue: row head
            AXRole: AXCell AXValue: 
                AXRole: AXStaticText AXValue: data
                    AXRole: AXInlineTextBox AXValue: data
            AXRole: AXRowHeader AXValue: 
                AXRole: AXStaticText AXValue: row head
                    AXRole: AXInlineTextBox AXValue: row head
    AXRole: AXTable AXValue: 
        AXRole: AXRow AXValue: 
            AXRole: AXRowHeader AXValue: 
                AXRole: AXStaticText AXValue: row head
                    AXRole: AXInlineTextBox AXValue: row head
            AXRole: AXCell AXValue: 
                AXRole: AXStaticText AXValue: data
                    AXRole: AXInlineTextBox AXValue: data
            AXRole: AXCell AXValue: 
                AXRole: AXStaticText AXValue: data
                    AXRole: AXInlineTextBox AXValue: data
        AXRole: AXRow AXValue: 
            AXRole: AXColumnHeader AXValue: 
                AXRole: AXStaticText AXValue: column head
                    AXRole: AXInlineTextBox AXValue: column head
            AXRole: AXColumnHeader AXValue: 
                AXRole: AXStaticText AXValue: column head
                    AXRole: AXInlineTextBox AXValue: column head
            AXRole: AXColumnHeader AXValue: 
                AXRole: AXStaticText AXValue: column head
                    AXRole: AXInlineTextBox AXValue: column head
        AXRole: AXRow AXValue: 
            AXRole: AXCell AXValue: 
                AXRole: AXStaticText AXValue: data
                    AXRole: AXInlineTextBox AXValue: data
            AXRole: AXCell AXValue: 
                AXRole: AXStaticText AXValue: data
                    AXRole: AXInlineTextBox AXValue: data
            AXRole: AXCell AXValue: 
                AXRole: AXStaticText AXValue: data
                    AXRole: AXInlineTextBox AXValue: data
    AXRole: AXParagraph AXValue: 
        AXRole: AXStaticText AXValue: End of test
PASS successfullyParsed is true

TEST COMPLETE
